Refactor email models in order to not create abstracts in the database
