+++ Apps/Emails/admin.py
@@ -1,11 +1,11 @@
 from django.contrib import admin
 from django.db.models import Model
 
-from Emails.models.models import BlackList
-from Emails.models.models import Block
-from Emails.models.models import Email
-from Emails.models.models import Notification
-from Emails.models.models import Suggestion
+from Emails.models import BlackList
+from Emails.models import Block
+from Emails.models import Email
+from Emails.models import Notification
+from Emails.models import Suggestion
 
 
 class BlockAdmin(admin.ModelAdmin):
