+++ Apps/Emails/factories/suggestion.py
@@ -5,8 +5,8 @@ from rest_framework.exceptions import ParseError
 
 from Emails.choices import CommentType
 from Emails.factories.block import SuggestionBlockFactory
-from Emails.models.models import Block
-from Emails.models.models import Suggestion
+from Emails.models import Block
+from Emails.models import Suggestion
 
 
 class SuggestionEmailFactory(factory.django.DjangoModelFactory):
