+++ Apps/Emails/tests/test_models.py
@@ -14,12 +14,11 @@ from Emails.factories.email import EmailFactory
 from Emails.factories.notification import NotificationFactory
 from Emails.factories.suggestion import SuggestionEmailFactory
 from Emails.fakers.suggestion import SuggestionErrorFaker
-from Emails.models.abstracts import AbstractEmailClass
-from Emails.models.models import BlackList
-from Emails.models.models import Block
-from Emails.models.models import Email
-from Emails.models.models import Notification
-from Emails.models.models import Suggestion
+from Emails.models import BlackList
+from Emails.models import Block
+from Emails.models import Email
+from Emails.models import Notification
+from Emails.models import Suggestion
 from Users.fakers.user import EmailTestUserFaker
 from Users.fakers.user import UserFaker
 from Users.models import User
@@ -45,10 +44,6 @@ class TestBlockModel:
 
 @pytest.mark.django_db
 class TestEmailModel:
-    def test_get_emails_abstract_class_fails(self) -> None:
-        abstract: AbstractEmailClass = AbstractEmailClass()
-        with pytest.raises(ValueError):
-            abstract.get_email()
 
     def test_email_attributes(self) -> None:
         email: Email = EmailFactory()
