+++ Apps/Emails/tests/test_permissions.py
@@ -4,7 +4,7 @@ from mock import MagicMock
 from mock import PropertyMock
 
 from Emails.fakers.blacklist import BlackListTestFaker
-from Emails.models.models import BlackList
+from Emails.models import BlackList
 from Emails.permissions import IsBlacklistOwner
 from Users.fakers.user import AdminFaker
 from Users.fakers.user import UserFaker
