+++ Apps/Emails/tests/test_utils.py
@@ -2,7 +2,7 @@ import pytest
 from django.core import mail
 from django_rest_passwordreset.models import ResetPasswordToken
 
-from Emails.models.models import Email
+from Emails.models import Email
 from Emails.utils import send_email
 from Users.fakers.user import UserFaker
 from Users.models import User
