+++ Apps/Emails/utils.py
@@ -2,7 +2,7 @@ from django_rest_passwordreset.models import ResetPasswordToken
 
 from Emails.factories.email import ResetEmailFactory
 from Emails.factories.email import VerifyEmailFactory
-from Emails.models.models import Email
+from Emails.models import Email
 from Project.utils.log import log_email_action
 from Users.models import User
